This commit was manufactured by cvs2svn to create tag '1.8.3'.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/tags/1.8.3@8297 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/import-export/qif-import/druid-qif-import.c
@@ -675,7 +675,7 @@ gnc_ui_qif_import_select_loaded_file_cb(GtkCList   * list,
   QIFImportWindow * wind = user_data;
 
   if(gh_list_p(wind->imported_files) && 
-     (gh_length(wind->imported_files) > row)) {
+     ((int)gh_length(wind->imported_files) > row)) {
     scm_unprotect_object(wind->selected_file);
     wind->selected_file = gh_list_ref(wind->imported_files,
                                       gh_int2scm(row));
